# src/mcp/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Task.sh MCP Server",
    "description": "Model Context Protocol server for Task.sh - Provides task context and export tools for LLMs",
    "version": "1.0.0",
    "contact": {
      "name": "Task.sh API Support"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3002",
      "description": "Development MCP Server"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API_KEY",
        "description": "API key for authenticating with the MCP server"
      }
    },
    "schemas": {
...
      "ContextResponse": {
        "type": "object",
        "properties": {
          "success": {"type": "boolean"},
          "context_type": {"type": "string"},
          "user_id": {"type": "string"},
          "timestamp": {"type": "string", "format": "date-time"},
          "data": {"type": "array"},
          "summary": {"type": "object"}
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {"type": "boolean", "enum": [false]},
          "error": {"type": "string"},
          "code": {"type": "string"},
          "message": {"type": "string"}
        }
      }
    }
  },
  "paths": {
...
    "/context/tasks/all": {
      "get": {
        "summary": "Get all tasks",
        "description": "Retrieves all tasks (complete and incomplete) for the authenticated user",
        "responses": {
          "200": {
            "description": "All tasks retrieved successfully",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ContextResponse"}
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ErrorResponse"}
              }
            }
          }
        }
      }
    },
...
    "/context/tasks/search": {
      "get": {
        "summary": "Search tasks",
        "description": "Searches tasks by description for a given keyword",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "description": "The keyword to search for",
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "Search results retrieved successfully",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ContextResponse"}
              }
            }
          },
          "400": {
            "description": "Invalid search query",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ErrorResponse"}
              }
            }
          }
        }
      }
    },
    "/context/tasks/{taskId}": {
      "get": {
        "summary": "Get specific task",
        "description": "Retrieves detailed information for a specific task by its ID",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "description": "The unique ID of the task",
            "schema": {"type": "string"}
          }
        ],
        "responses": {
          "200": {
            "description": "Task retrieved successfully",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ContextResponse"}
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ErrorResponse"}
              }
            }
          }
        }
      }
    },
...
  },
...
}